feat: add the macro for some extension field(Fp12, Fq12) tests
