+++ src/tests/field.rs
@@ -678,6 +678,88 @@ macro_rules! field_testing_suite {
         }
     };
 
+    ($ext_field: ident, "f12_tests", $base_field_1: ident, $base_field_2: ident) => {
+        #[test]
+        fn test_f12_mul_by_014() {
+            let mut rng = XorShiftRng::from_seed([
+                0x59, 0x62, 0xbe, 0x5d, 0x76, 0x3d, 0x31, 0x8d, 0x17, 0xdb, 0x37, 0x32, 0x54, 0x06, 0xbc,
+                0xe5,
+            ]);
+
+            for _ in 0..1000 {
+                let c0 = $base_field_2::random(&mut rng);
+                let c1 = $base_field_2::random(&mut rng);
+                let c5 = $base_field_2::random(&mut rng);
+                let mut a = $ext_field::random(&mut rng);
+                let mut b = a;
+
+                a.mul_by_014(&c0, &c1, &c5);
+                b.mul_assign(&$ext_field {
+                    c0: $base_field_1 {
+                        c0,
+                        c1,
+                        c2: $base_field_2::zero(),
+                    },
+                    c1: $base_field_1 {
+                        c0: $base_field_2::zero(),
+                        c1: c5,
+                        c2: $base_field_2::zero(),
+                    },
+                });
+
+                assert_eq!(a, b);
+            }
+        }
+
+        #[test]
+        fn test_f12_mul_by_034() {
+            let mut rng = XorShiftRng::from_seed([
+                0x59, 0x62, 0xbe, 0x5d, 0x76, 0x3d, 0x31, 0x8d, 0x17, 0xdb, 0x37, 0x32, 0x54, 0x06, 0xbc,
+                0xe5,
+            ]);
+
+            for _ in 0..1000 {
+                let c0 = $base_field_2::random(&mut rng);
+                let c3 = $base_field_2::random(&mut rng);
+                let c4 = $base_field_2::random(&mut rng);
+                let mut a = $ext_field::random(&mut rng);
+                let mut b = a;
+
+                a.mul_by_034(&c0, &c3, &c4);
+                b.mul_assign(&$ext_field {
+                    c0: $base_field_1 {
+                        c0,
+                        c1: $base_field_2::zero(),
+                        c2: $base_field_2::zero(),
+                    },
+                    c1: $base_field_1 {
+                        c0: c3,
+                        c1: c4,
+                        c2: $base_field_2::zero(),
+                    },
+                });
+
+                assert_eq!(a, b);
+            }
+        }
+
+        #[test]
+        fn test_squaring() {
+            let mut rng = XorShiftRng::from_seed([
+                0x59, 0x62, 0xbe, 0x5d, 0x76, 0x3d, 0x31, 0x8d, 0x17, 0xdb, 0x37, 0x32, 0x54, 0x06, 0xbc,
+                0xe5,
+            ]);
+
+            for _ in 0..1000 {
+                let mut a = $ext_field::random(&mut rng);
+                let mut b = a;
+                b.mul_assign(&a);
+                a.square_assign();
+                assert_eq!(a, b);
+            }
+        }
+    };
+
     ($ext_field: ident, "frobenius", $frobenius_param: expr) => {
         #[test]
         fn test_frobenius() {
